Reorder and add content
Reorders current documentation about building raw images and ISOs.
It splits the 'Creating derivatives' docs in proper section which are
general for other cases as well, and move it back as a simple example.

More emphasis is provided on building simply docker images that can be
used by the tooling, so the Build ISOs section has been restructured,
and the Packer section has been created too.

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>