Generate docs from specfiles
This PR introduce a script which generates layout/structure to be parsed
from our docs to be displayed in the website. It parses the cos-toolkit
repository packages and scans for `README.md` files present in the
package path (where `build.yaml` and `definition.yaml` reside).

If it's a collection, it still reads for README files, but allows also
to generate per-package documentation by checking the presence of
`README-cat-name.md` in the package path, if present it will create a
new grouped page and put the packages of the collection under it.

In order to preserve readability on cos-toolkit side, the page header of
a collection cannot be generated cleanly, so a `README-head.md` file is
required with the specific docsy fields, e.g:

```yaml
---
title: "Collection name"
linkTitle: "Collection title"
date: 2021-12-02
description: >
  Collection description
---
```

Needs: https://github.com/rancher-sandbox/cOS-toolkit/pull/907
Part of: https://github.com/rancher-sandbox/cOS-toolkit/issues/757

Add makefile target and hook into CI

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>